templates/python/requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
uvicorn==0.15.0
cchardet==2.1.7
pyserde[msgpack]==0.8.3
pylru==1.2.0
